Refactor Result and Result<TValue>- moved members to extensions
Most members of Result and Result<TValue> have
been moved to extension method classes.

Here we followed the following rule, input
variables are as unprecise as possible, so
IResultBase instead of result, but output
types are as precise as possible, so Result<TValue>
instead of IResult<TValue>.

We had to double some of the extension methods,
because Task and ValueTask are not covariant,
we had to provider methods with Task<Result<T>>
and alos with Task<IResult<T>> etc.

I added IResultWithoutValue, so that we can
write extensions methods, that are only valid
for Result, but not for Result<T>.